Bug14762796 was first fixed in rb:1416, reviewed and approved by Jimmy.
But that fix showed several test cases failing intermittently on PB2.
CreateThread() does not immediately start the new thread so that in
some small amount of calls to os_thread_create(), the CloseHandle()
added by the previous patch actually closed the thread object before
the thread itself could open a handle to it.  Then when the thread
actually started, some of that thread info was lost.  This solution
of closing the thread handle as soon as the thread was created was
found in the MSDN documentation as a comment to the official docs.
It never mentioned being timing dependent.

I also noted that contrary to the MSDN documentation, not calling
ThreadExit() from the thread itself before exiting in C++ code
caused certain testcases to fail because of unreleased resources,
at least some of the time on Windows 32-bit Vista.  I could never
reproduce either problem on Windows 7-64.

So I pushed a small patch to mysql-trunk to fix the failing
testcases by adding a Sleep(0) before the CloseHandle() in
os_thread_create() and to go back to ALWAYS doing ThreadExit() in
os_thread_exit().  This seems to have cleaned up the failing testcases.

But that 'Sleep(0)' solution is too indeterminate.  This patch
introduces a global stl::map object to remember the Windows Handles
returned from CreateThread() and then close them in os_thread_exit().
This way they are not closed immediately while the thread is still
getting started, which relys on thread switching and timing to work
correctly.

In addition, this patch makes sure that the Linux code always returns
the new_thread_id in the thread_id parameter to os_thread_create().

Approved by Sunny in rb:1788