Bug#16094171: UPDATE_CONST_EQUAL_ITEMS() ADDS INDEXED FIELD               TO JT::CONST_KEYS BUT NOT JT::KEYS
When update_const_equal_items() finds that a field is constant
due to multiple equality, indexes that field is part of are
added to the JOIN_TAB::const_keys. However, the same indexes
should also be added to JOIN_TAB::keys. Because this was not
done, a DBUG_ASSERT that checks that const_keys is a subset
of keys failed.