Bug#14626800 ASSERTION FAILURE IN DEBUG_SYNC()
MySQL has debug sync facility that is enabled by default 
on PB2 machines. This facility is used by storage engines
 to allow synchronization points(which are usually used 
by test code).

When a connection is closed, the following things are done 
during connection cleanup in the order given below:

a) Cleanup is done for the debug sync facility.
b) Temporary tables are closed. 

While the temporary tables are being closed by storage
engine, storage engine tries to use the debug sync facility.
However since debug sync facility is already cleaned up,
an assert is hit in debug sync.

The fix is to do the cleanup related to debug sync after the 
temporary tables are closed. We are not removing the assert 
causing the failure because doing so can mask the bug when 
debug sync points are not behaving as intended.