BUG#14040155 - FORMAT FUNCTION WITH LOCALE CRASHES WITH BIG NUMBERS
Problem:
When we are using modulo operation with big number inside FORMAT function the 
mysql daemon crashes due to this assertion statement.
      
DBUG_ASSERT(dec_length <= str_length);
      
dec_length-Decimal places we want.
str_length-string length of the resultant string.
      
      
Analysis:
We save our decimal data in mydecimal class which has member decimal_digit_t 
buffer[DECIMAL_BUFF_LENGTH] to store decimal values. The decimal_digit_t is a 
signed integer and DECIMAL_BUFF_LENGTH is macro of 9.
      
Single decimal_digit_t can carry at most 9 'decimal' digit value(that is it 
can take value less then  999,999,999) and we have 9 decimal_digit_t which 
mean we have space for 81 decimal digit max(that is  max value it can hold 
is 81 times repeated '9') and when we have to define a decimal digits after
point they have to allocate space of multiple of 9.
      
So that means if we have value like 12345123456789.1234
      
it will save it like this
      
12345        123456789        123400000
buf[0]        buf[1]            buf[2]
      
-->Integer part is save in buf[0] and buf[1] with buf have max 9 digits 
and fraction part must carry 9 digits(like in fraction we have 1234 and we change it to 
123400000 in buf[2]).
      
suppose we have a query like this-->
select format((72 times repeated '9')%(72 times repeated '9'),27,'pl_PL');
      
In this case we have expression((72 times repeated '9')%(72 times repeated '9')) 
whose value should be rounded to 27 decimals.
      
We save our expression in decimal_digit_t. When we do module of this 
expression, the result will have 72 times repeated '0',then the space available
for fraction part is of 9 decimal digit, so we truncate the decimal digit(that 
is given to us as a second argument in format function in the query ) from 27 
to 9, as that much is available to us.
After that we calculate the length of the integer part (0), so the 
str_length will be 0+9(truncated value of fraction part) and 
dec_length(27), result in debug assertion fail as string length(9) is 
less then decimal length(27).
      
example of module operation->
      
(71 times repeated '9')%(70 times repeated '9'+'8')
      
numerator buffer will be store like this.
      
99999999	999999999	999999999	999999999	999999999	999999999	999999999	999999999	some garbage	
buf[0]		buf[1]		buf[2]		buf[3]		buf[4]		buf[5]		buf[6]		buf[7]		buf[8]		
      
denominator buffer will be store like this.
      
99999999	999999999	999999999	999999999	999999999	999999999	999999999	999999998	some garbage	
buf[0]		buf[1]		buf[2]		buf[3]		buf[4]		buf[5]		buf[6]		buf[7]		buf[8]		
      
      
the result of this module will be in numerator then store to resultant buffer.
the resultant numerator may carry sumthing like this
      
somethg nt usable	0		0		0		0		0		0		1		some garbage	
buf[0]			buf[1]		buf[2]		buf[3]		buf[4]		buf[5]		buf[6]		buf[7]		buf[8]		
      
from this numerator we store the result in resultant buffer.
So result in resultant buffer will be 
0		0		0		0		0		0		1		some garbage	some garbage	
buf[0]		buf[1]		buf[2]		buf[3]		buf[4]		buf[5]		buf[6]		buf[7]		buf[8]		
      
Now as we see, we have space for 2 buf(that is buf[8] ) so we truncate our 
decimal value from 27 to 18.
      
The length that will be 18(truncated value)+
1(for decimal point)+1(moduler result that is 1).
      
Which cause our debug assert to fail.
      
DBUG_ASSERT(dec_length(27) <= str_length(18));
where dec_length is the dec_length which we want while running the 
query and str_length is string length of resultant query.
       
      
Solution:
      
We remove the leading zeroes from the result of the division operator, 
so that we can get actual length of the result.Which will help us to 
decide whether we need to truncate the decimal part of the result.