Bug#14678248: SLAVE_OPEN_TEMP_TABLES_LOCK INITIALIZED MULTIPLE TIMES
The slave_open_temp_tables_lock can currently be initialized
multiple times because this action is made on the Relay_log_info
constructor. In fact, this constructor is not only used by the
SQL thread itself but also by clients to execute Binlog
statements where dummy relay log info objects are needed. Not
belonging to the class, the double initialization of this lock
can be problematic.

Guaranteeing that the lock is only initialized on the relay log
info object that belongs to the SQL thread solves this issue.