rb://1611 Approved by Jimmy and Marko.
In order to implement GIS indexing in InnoDB, InnoDB needs to support
GEOMETRY datatypes, so that InnoDB stores them properly and understands
them.  Currently, InnoDB store all geometry types in BLOB. We can
continue to use BLOB (other than point) as underlying datatype, but the
column needs to be marked as proper datatypes.