Backport trunk => 5.6
Fix for "Thread stack overrun" on 64-bit sparc in DBUG mode.
Stack frames are quite big: 7648 bytes as computed by used_stack()
The size shrinks to 1616 bytes if we switch on '-xO2' or more,
but we don't want to that for DBUG builds.

So we change the problematic query instead.