Bug#16088188  'STOP SLAVE' MAY CAUSE DEADLOCK Being stopped the slave IO thread can deadlock with one that queries SHOW STATUS LIKE 'heartbeat_period'.
Reason of the deadlock resembles bug 14236151 case.
The common part is the contative LOCK_active_mi.
A difference is LOCK_status in place of LOCK_global_system_variables.
Ordering of LOCK_active_mi and LOCK_status aquisition must be
as they are listed in this clause.

Fixed with applying a similar to bug 14236151 pattern 
to temporarily release the critical participant of deadlock
inside a status function.
It's LOCK_status this time.

The following set of status is found out to be vulnerable:

  {"Slave_retried_transactions",(char*) &show_slave_retried_trans, SHOW_FUNC},
  {"Slave_heartbeat_period",   (char*) &show_heartbeat_period, SHOW_FUNC},
  {"Slave_received_heartbeats",(char*) &show_slave_received_heartbeats, SHOW_FUNC},
  {"Slave_last_heartbeat",     (char*) &show_slave_last_heartbeat, SHOW_FUNC},
  {"Slave_running",            (char*) &show_slave_running,     SHOW_FUNC},

Their functions are fixed.