Bug#1624555 - PATCH FOR BUG#14729365 CAUSE RO PERFORMANCE REGRESSION FOR 8 CORE HOST(S)
Use a system specific random indexer in the fuzzy counter class. This helps in
reducing the cache coherency impact when multiple threads try and update
the same counter.

Note: This patch uses THD::thread_id if a "fast" random indexer is not available
for that OS. However, for srv_stats.n_rows_read this "fast" part is moot because
a static thread specific identifier should be faster than a function call. Given
that the performance tests were done with the OS specific indexer I've decided
to leave that as is for now. We should revisit this.

Approved by Jimmy Yang, rb#1950