This patch fixes the bug#13875572 - NEXT_INSERT_ID >=                                     AUTO_INC_INTERVAL_FOR_CUR_ROW.MINIMUM
This bug could led to a crash for the server built in debug mode.

There was a bug in the control of overflow for the value of variable
 next_insert_id. Due to lack of overflow checking it was possible for
next_insert_id variable in handler::update_auto_increment() method
to get value greater than maximum value of column.
In such a situation next attempt to generate auto-increment by calling
ha_innobase::get_auto_increment() method led to allocation of
auto-increment interval which covered some values already used
(since InnoDB interpreted this as attempt to insert negative value).
Future attempts to use values from this interval might led to duplicate
key errors. In such a case we might have allocated new auto-increment
interval, but due to IGNORE duplicate key handling mode
handler::restore_auto_increment() was called which restored next_insert_id
to value lower than lower border of freshly allocated
interval. As result on next call to update_auto_increment() we got
assertion failure mentioned above.

As a side effect of this patch the server will adjust next_insert_id only
in the case when the value specified in the INSERT is positive. It means
that as of now we don't treat negative auto-increment values as large positive
values. This is incompatible change and should be documented.