bug#16079840  PREVIOUS_GTIDS_LOG_EVENT SHOULD HAVE EXTENSIBLE DATA SECTION
The Previous_gtids_log_event currently does not allow adding data at the
end of the event: if extra data is found, an error is generated.

Fix to make Previous_gtids_log_event work fine with additional unknown
data at the end of the event, then that will be silently ignored.