Bug#16407965: Item::save_in_field_no_warning() does not return correct conversion               status for strings
When using Item::save_in_field_no_warning() on string data types, it always
returns TYPE_OK even when the string is truncated due to being longer than
the field it is stored into. Due to this, Item::save_in_field_no_warning()
can not be used to determine if a string value fits completely into a string
field or not.

This patch changes Item::save_in_field_no_warning() for string items so that
TYPE_NOTE_TRUNCATED or TYPE_WARN_TRUNCATED is returned if the string does
not fit in the field. 

This patch also contains a few additional changes (in the range optimizer
and in the code for execution of subqueries) that used 
Item::save_in_field()/Item::save_in_field_no_warning() but did not handle
that TYPE_WARN_TRUNCATED got returned.

A unit test for Item_string::save_in_field() and 
Item_string::save_in_field_no_warning() is added. This shows that without this
change Item_string::save_in_field_no_warning() returns TYPE_OK even when
a longer string than the field can store is saved in the field.