Bug#13641256: 5.5.16 WAITS FOR THE QUERY CACHE               EVEN WITH SQL_NO_CACHE HINT
The problem was that the SQL_NO_CACHE clause for SELECT
didn't always work correctly and caused query cache
mutexes to be locked even if the query was not supposed
to use query cache.

The problem was that the string matching used to check
for the presence of SQL_NO_CACHE assumed that it would
have a space char before and after. So if the query was
written with e.g. a line shift before and/or after SQL_NO_CACHE
the pattern would not match and SQL_NO_CACHE would not be effective.

Note that check occurs before proper SQL parsing. A query
cache hit is meant to avoid the overhead of proper parsing.

This patch fixes the problem by checking that there is at
least one whitespace character before and after SQL_NO_CACHE
instead of checking for the presence of space.